Create knights_travails form Odin Project
I created a method knight_moves which determines the fastest way for a
chess knight to get from a starting point to a target point.
I first tried to create a data structure with interconnected nodes, but
it failed to traverse. Then I tried again creating a graph structure,
which worked properly.
For traversing the graph I used a BFS algorithm to check whether a node
is the target. If it is, the algorithm and returns the path from start
to there and if not adds all the neighbors of the node to the queue.